Updated code based on name change from "sparse transformer" to "sparse attention" (#76)
* sparse attention name change

* updated config, setup, and tests
