+++ deepspeed/ops/sparse_attention/trsrc/softmax_bwd.tr
@@ -1,4 +1,4 @@
-// DeepSpeed note, code taken & adapted from commit beddcc3eda6d3df1b34f74c2e10139e4317d0e7f
+// DeepSpeed note, code taken & adapted from commit 9aa94789f13ada713af36cfd8cca2fc9a7f6b79a
 // https://github.com/ptillet/torch-blocksparse/blob/master/torch_blocksparse/softmax.py
 
 __global__ void softmax_bwd(TYPE * X __readonly __noalias __aligned(16),
